web/requirements.txt
Django
celery
redis
librabbitmq
djangorestframework
nltk
futures; python_version < '3.0'